Paging now supports 4KB pages as well as 2MB pages

+++ graphics/graphics.c
@@ -36,7 +36,7 @@ void graphics_Initialize()
         frameBufferB = bootstrap_malloc(buffer_size);
 
         char *vPointer = virtMemMan_FindEmptyAddress(buffer_size, MEM_KERNEL);
-        int retVal = virtMemMan_Map(vPointer, frameBufferA, buffer_size, MEM_TYPE_WC, MEM_WRITE, MEM_KERNEL);
+        int retVal = virtMemMan_Map((uint32_t)vPointer, (uint64_t)frameBufferA, buffer_size, MEM_TYPE_WC, MEM_WRITE, MEM_KERNEL);
         frameBufferA = vPointer;
 
         //frameBufferB = bbuf;
